Extracting IO into its own file
in preparation for porting buffered io bits from pypy, mostly.  Also, it
took me awhile to find W_IOObject in the first place because I was
looking for a file with "io" in the name.
